2024-06-26
Long day,  I think I have the training working properly

seems to crash

Prof wants to have mode evaluated on nodes,
I am assuming that this is done by making a grid
in the xy-plane and in time.

to accomplish this, needing to add functionability to
cpp file, make class and call functions.